Initial commit for recoding: database.py is rewritten
I have decided to rewrite this project, since I generally include it
in my resume, but the code is too clumsily written. My purpose is to
remove the unnecessary stuff, and writing it in a way that will comply
with conventions as much as possible.

I switched from sql alchemy to mysql-connector-python library, since I
believe using SQL syntax gives me more freedom, it lets the code be able
to do a job with one (or less) function instead of more. See what I did with
database.get_all, for instance. Instead of using dedicated functions
like get_all_countries and get_user_ids, which are basically the same
code with different column/table names, I use get_all as a sql query
wrapper which anything can be fetched with via indicating table and column
names as a parameter.