Data are fetched at once instead of by country.
There are changes in how updates work. Instead of fetching by country,
all are requested at once from "https://corona.lmao.ninja/v2/countries",
and a dictionary is created. This prevents getting errors from the
server, and not making more than 100 of requests in every 15 minutes or so is more respectful to the owner of the API. Also, the speed of updates are significantly increased.